Add new Kafka Producer config property (#21)
This adds a new possibility when setting up a Kafka exporter: specifying a value for the producer property message.max.bytes. This will allow us to send larger messages to Kafka.

The value must be an integer representing the maximum size of the message in bytes. To specify a value, we just need to add a new key "message_max_bytes" to the config array passed to the constructor of the class Arquivei\Events\Sender\Exporters\Kafka.